Refactor shared diagram direction handling
