+++ src/diagrams/er/types.rs
@@ -2,6 +2,8 @@
 
 use std::collections::HashMap;
 
+pub use crate::diagrams::direction::Direction;
+
 /// Cardinality types for ER relationships
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
 pub enum Cardinality {
@@ -196,37 +198,6 @@ impl EntityClass {
             id,
             styles: Vec::new(),
             text_styles: Vec::new(),
-        }
-    }
-}
-
-/// Direction of the ER diagram
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub enum Direction {
-    #[default]
-    TopToBottom,
-    BottomToTop,
-    LeftToRight,
-    RightToLeft,
-}
-
-impl Direction {
-    #[allow(clippy::should_implement_trait)]
-    pub fn from_str(s: &str) -> Self {
-        match s.to_uppercase().as_str() {
-            "BT" => Self::BottomToTop,
-            "LR" => Self::LeftToRight,
-            "RL" => Self::RightToLeft,
-            _ => Self::TopToBottom,
-        }
-    }
-
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            Self::TopToBottom => "TB",
-            Self::BottomToTop => "BT",
-            Self::LeftToRight => "LR",
-            Self::RightToLeft => "RL",
         }
     }
 }
